+++ db/migrate/20110301202016_create_calibration_levels.rb
@@ -2,7 +2,7 @@ class CreateCalibrationLevels < ActiveRecord::Migration
   def self.up
     create_table :calibration_levels do |t|
       t.integer :run_log_id
-      t.integer :standard_conc
+      t.float :standard_conc
 
       t.timestamps
     end
